+++ Modules/core/CoreRouting.php
@@ -26,7 +26,8 @@ class CoreRouting extends Routing{
         
         // create account
         $this->addRoute("corecreateaccount", "corecreateaccount", "coreaccount", "index");
-        $this->addRoute("coreaccountcreated", "coreaccountcreated", "coreaccount", "created");         
+        $this->addRoute("coreaccountcreated", "coreaccountcreated", "coreaccount", "created");
+        $this->addRoute("coreaccountisunique", "coreaccountisunique", "coreaccount", "isunique", array("type", "value", "id_user"), array("", "", ""));
          
         // home
         $this->addRoute("corehome", "corehome", "corehome", "index");
@@ -47,7 +48,6 @@ class CoreRouting extends Routing{
         $this->addRoute("coreusersedit", "coreusersedit", "coreusers", "edit", array("id"), array(""));
         $this->addRoute("coreusersdelete", "coreusersdelete", "coreusers", "delete", array("id"), array(""));
         $this->addRoute("coreuserslanguageedit", "coreuserslanguageedit", "coreusers", "languageedit");
-        $this->addRoute("coreusersisunique", "coreusersisunique", "coreusers", "isunique", array("type", "value", "id_user"), array("", "", ""));
 
         // settings
         $this->addRoute("coremyaccount", "coremyaccount", "coreusers", "myaccount");