+++ cli/pfm-cli.php
@@ -3,6 +3,7 @@ require __DIR__ . '/../vendor/autoload.php';
 
 require_once 'Framework/Configuration.php';
 require_once 'Framework/FCache.php';
+require_once 'Framework/Router.php';
 
 require_once 'Modules/core/Model/CoreInstall.php';
 require_once 'Modules/core/Model/CoreUser.php';
@@ -31,6 +32,9 @@ $cli = Cli::create()
     ->command('version')
     ->description('Show version')
     ->opt('db:d', 'Show installed and expected db version', false, 'boolean')
+    ->command('cache')
+    ->opt('clear', 'Clear caches', false, 'boolean')
+    ->opt('dry', 'Dry run', false, 'boolean')
     ->command('repair')
     ->opt('bug', 'Bug number', 0, 'integer');
 
@@ -49,6 +53,22 @@ try {
                 $modelCache = new FCache();
                 $modelCache->freeTableURL();
                 $modelCache->load();
+            } else {
+                $r = new Router();
+                $rl = $r->listRoutes();
+                echo "Routes:\n";
+                foreach ($rl as $rll) {
+                    echo "* $rll[0] - $rll[1]\n";
+                }
+                $fc = new FCache();
+                $rl = $fc->listAll();
+                foreach ($rl as $rll) {
+                    $url = "* GET|POST - $rll[0] /$rll[0]";
+                    for($i=1;$i<count($rll);$i++) {
+                        $url .= "/[i:".$rll[$i]."]";
+                    }
+                    echo "$url\n";
+                }
             }
             break;
         case 'expire':
@@ -69,11 +89,44 @@ try {
                 echo "DB expected version: ".$cdb->getVersion()."\n";
             }
             break;
+        case 'cache':
+            if($args->getOpt('clear')) {
+                cacheClear($args->getOpt('dry'));
+            }
+            break;
         default:
             break;
     }
 } catch(Throwable $e) {
     Configuration::getLogger()->error('Something went wrong', ['error' => $e->getMessage(), 'stack' => $e->getTraceAsString()]);
+}
+
+function cacheClear($dry) {
+    if(is_dir('/tmp/pfm')) {
+       removeDirectory('/tmp/pfm', $dry);
+       Configuration::getLogger()->info('cache cleared');
+    } else {
+        Configuration::getLogger()->info('nothing to do');
+    }
+}
+
+function removeFile($path, $dry=false) {
+    if($dry) {
+        Configuration::getLogger()->info('[delete]', ['path' => $path]);
+    } else {
+	    unlink($path);
+    }
+}
+function removeDirectory($path, $dry=false) {
+	$files = glob($path . '/*');
+	foreach ($files as $file) {
+		is_dir($file) ? removeDirectory($file, $dry) : removeFile($file, $dry);
+	}
+    if($dry) {
+        Configuration::getLogger()->info('[delete]', ['path' => $path]);
+    } else {
+	    rmdir($path);
+    }
 }
 
 function cliFix($bug=0) {